DPSPS-50: Initial implementation of DBCursor working properly
This changeset is part of the refactor away from MongoCasbah to modern Mongo drivers whilst providing a backwards compatible shim for the MongoDBObject and CasbahRepository interfaces we use. It is a large piece of work which is still in progress.

MongoCasbah oftens throws Exceptions and we maintain this behaviour for compatibility, even though we prefer the use of Either.

The major change in this update is that DBCursor now delays execution of queries, allowing find to be chained with sort, limit, skip etc. So this:

    repo.find(...).projection(...).sort(...).skip(...).limit(...).map(...)

Only executes one DB query, and all the commands are executed server side as expected.

map() is lazily chained onto dbcursor so dbCursor.map(ab).map(bc).map(cd) only iterates over the realised resultset once, applying (cd(bc(ab(_)))) once, rather than turning an entire List[A] -> List[B] -> List[C]. This is vastly more efficent.

Other key bugfixes in this changeset:

  * getInsertedId() and getUpsertedId() fixed on CasbahInsertResult and CasbahWriteResult
  * isUpdateOfExisting() fixed
  * Fixed encoding of Double, Float and BigDecimal in MongoDBObjects
  * Throw an error instead of stringify unknown primitive types
  * Fix nesting of MongoDBObjects correctly, so that Maps don't leak out of equality checks.
  * casbahRepo.find(filter, projection) now actually includes the projection instead of ignoring it
  * New tests added for MongoCasbahRepository